pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "matchkit"
version = "1.0.0"
description = "Open-source, white-label member engagement and matching platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["matching", "member-engagement", "white-label", "fastapi", "embeddings"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Data validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "email-validator>=2.1.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    # AI / Embeddings
    "openai>=1.30.0",
    "numpy>=1.26.0",
    # Email (Mandrill + SendGrid fallback)
    "mailchimp-transactional>=1.0.56",
    "sendgrid>=6.11.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # Templates
    "jinja2>=3.1.0",
    # Data loading
    "pandas>=2.2.0",
    # HTTP client (CRM integration)
    "httpx>=0.27.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Config
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
include = [
    "api*", "config*", "crm*", "db*", "email_service*",
    "matching*", "models*", "scheduler*", "utils*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]

[tool.ruff]
target-version = "py311"
line-length = 100
